Fix dynamic for_each error for archs
```
Error: Invalid for_each argument

  on .terraform/modules/base/main.tf line 55, in resource "cosign_attest" "this":
  55:   for_each = var.skip_attest ? [] : local.archs
    ├────────────────
    │ local.archs is a set of dynamic, known only after apply
    │ var.skip_attest is false
```

We can't derive this easily from the build output, even though it's the right thing to do.

This isn't an optimal solution, but this moves the arch value to a var, which then requires
that the archs must match the built archs else it will error on signing. I don't particularly
like this, but I don't know how else we can do this within only terraform without trying to determin
this based on computed dynamic outputs. :(